+++ src/control_plane/gateway/mod.rs
@@ -1,0 +1,12 @@
+#[cfg(test)]
+mod tests;
+
+mod gateway_configuration;
+mod gateway_error;
+mod gateway_route;
+mod gateway_snapshot;
+
+pub(crate) use gateway_configuration::{GatewayConfiguration, GatewayFuture};
+pub(crate) use gateway_error::GatewayError;
+pub(crate) use gateway_route::GatewayRoute;
+pub(crate) use gateway_snapshot::GatewaySnapshot;